requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# File Storage
boto3==1.34.34
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-decouple==3.8
Pillow==10.2.0
python-dateutil==2.8.2

# SMS (Twilio)
twilio==8.13.0

# Development
django-extensions==3.2.3
ipython==8.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0